Adding Serializer Annotations to Link Model
Adding the annotations prevents the Serializer from deciding to exclude
the link when serializing. This can occur as a result of having used the
exclusionpolicy annotation on an entity that is serialized. This policy
bleeds through and is used in the acceptance decision when no exclusion
policy has been explicitly stated on relations