LinkFactory checks if there is Relation metadata, not just metadata
The MetadataFactory returns an object no matter if links are found or not
This leads to the fact that the strict null check always fails, and thus
always an empty array of links is added to the object, possibly filled
with links if specified. The check on the count of relations fails only
when there are no relations specified, as it should.